refactor(research_tracker/activity_polling): promote to directory, split tests (#1266)
activity_polling.rs (461 LOC) is promoted to a directory module:
- src/research_tracker/activity_polling/mod.rs: 279 LOC (down from 461)
- src/research_tracker/activity_polling/tests.rs: new (~182 LOC)

Pure refactor: zero logic changes, all 13 tests pass.

Closes 7/70 of #1266 (this loop iteration), 16 PRs cumulative since #1285.
